meta: allow grub config to be overwritten
OE-core doesn't make it easy to swap out grub.cfg
There's a recipe grub-bootconf that appears to allow
it, but from test seems to get overridin when do_image_complete
is called. This is due to a function called do_bootimg in
the image-live.bbclass. Which executes build_efi_cfg which
makes the actual grub.cfg.

Commit:
	1. removes the 'live iso' IMAGE_FSTYPES as by
	   default adding those will include the image-live.bbclass
	2. Allows grub.cfg to be completely generated from customer
	   input string
	3. Extracts important portions of oe-core so that we can
	   populate the image boot directory

Signed-off-by: Vincent Davis Jr <vince@underview.tech>