udoo-minimal-init: mount newly create IMAGES partition
Prior the wic image was storage on the same
partition as the boot files.

Seperation allows for mounting of partition
with wic on it as read/write. This wasn't
possible before do to isohybrid.

Commit finds the partition label IMAGES
and mounts it before using bmaptools to
update the emmc block device.

Signed-off-by: Vincent Davis Jr <vince@underview.tech>